Remove QueueItem component
