--- frontend/src/app/testQueue/QueueItem.scss
@@ -1,150 +1,0 @@
-.prime-card-container {
-  .card-header {
-    display: flex;
-    justify-content: flex-start;
-    max-width: 80%;
-
-    .card-name {
-      font-size: 22px;
-      font-weight: 700;
-      color: #005ea2;
-      cursor: pointer;
-      text-decoration: none;
-      line-height: 1.15;
-    }
-
-    .card-dob {
-      margin-left: 24px;
-      font-size: 16px;
-      font-weight: 400;
-      color: #71767a;
-      padding-top: 6px;
-
-      .card-date {
-        font-weight: 700;
-      }
-    }
-  }
-
-  .card-correction {
-    background-color: #fef0c8;
-    font-size: 16px;
-    font-weight: 400;
-    padding: 0.5rem 0 0.5rem 15px;
-  }
-
-  .card-correction-label {
-    color: #ffbe2e;
-  }
-
-  .card-correction-input {
-    border: 1px solid #ffbe2e;
-  }
-
-  .card-test-input {
-    height: 34px;
-    margin-right: 12px;
-    max-width: 130px;
-    animation: 0.25s fadeIn;
-    animation-fill-mode: both;
-
-    @keyframes fadeIn {
-      0% {
-        opacity: 0;
-      }
-
-      100% {
-        visibility: visible;
-        opacity: 1;
-      }
-    }
-
-    &__error {
-      border-color: #d54309;
-    }
-  }
-
-  .card-dropdown {
-    select {
-      height: 34px;
-      padding-top: 4px;
-      padding-bottom: 4px;
-    }
-  }
-
-  input[type="date"]::-webkit-calendar-picker-indicator {
-    margin-left: 0;
-  }
-
-  .check-box-container {
-    display: flex;
-    align-items: center;
-    height: 34px;
-  }
-
-  .test-questionnaire-btn {
-    font-weight: 700 !important;
-    margin-bottom: 1em;
-  }
-
-  .flex-col-container {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    flex-wrap: wrap;
-  }
-
-  .test-date-time-container {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    flex-wrap: wrap;
-    margin-top: 8px;
-  }
-
-  .queue-item-error-message {
-    color: #d54309;
-    display: block;
-  }
-}
-
-.symptom-onset-date {
-  height: 40px;
-  width: 200px;
-  margin-top: 8px;
-}
-
-.multiplex-result-form {
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: space-between;
-  height: 100%;
-  min-width: 24.5rem;
-}
-
-.prime-test-result-submit {
-  align-items: center;
-
-  .usa-button.prime-multiplex-btn {
-    width: 100px;
-    margin: 0;
-  }
-
-  div.usa-form-group {
-    margin-top: 0;
-  }
-
-  .usa-tooltip {
-    margin-right: 7px;
-
-    .usa-tooltip__body {
-      right: 20px;
-
-      @media (max-width: 619px) {
-        right: -100px;
-        bottom: 8px;
-        width: calc(100% - 60px);
-      }
-    }
-  }
-}
